+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [CONSIDERING]
 
-- TODO: add scheduled lcb defense changes for build/raid times
-- TODO: add zombie raid protection during build time
+- add admin command to enable/disable schedule feature and return to default server state immediately?
+- add more info to readme
+- create and add social image to project and readme header
+- finish teaser video
+- update build/raid text to reflect start/end times and timezone reference... perhaps +06:00 reference since these use numbers (cvar-supported)? :)
+
+## [0.1.0] - 2023-03-05
+
+- add admin command to toggle debug logging
+- add admin commands to update start/end time and timezone
+- add buff to raid protection state with instructions
+- add mechanic to trigger raid protection
+- add particle effects and sound on warp
+- add scheduled lcb defense for build/raid times
+- adjust formatting/colors of build/raid/protect panels
+- prevent warp from happening for spectator